+++ rcloud.support/R/rcloud.support.R
@@ -333,7 +333,7 @@ rcloud.update.notebook <- function(id, content, is.current = TRUE) {
                 stop(paste0(.session$username, " is not a member of protection group ", group$id))
         }
         ## take the content and encrypt it
-        enc <- .encrypt.by.group(list(files=l), group$id)
+        enc <- rcloud.encrypt(list(files=l), group=group$id)
         ## update contains just the encrypted piece
         ## if this is a conversion, remove the unencrypted pieces
         cfiles <- if (!isTRUE(old$content$is.encrypted)) .zlist(names(rcloud.get.notebook(id, raw=TRUE)$content$files)) else list()
@@ -576,7 +576,7 @@ rcloud.record.cell.execution <- function(user = .session$username, json.string) 
 #      file=pathConf("data.root", "history", "main_log.txt"), append=TRUE)
 }
 
-rcloud.debug.level <- function() if (hasConf("debug")) getConf("debug") else 0L
+rcloud.debug.level <- function() if (hasConf("debug")) as.integer(getConf("debug")) else 0L
 
 ################################################################################
 # stars