+++ NEWS.md
@@ -13,6 +13,11 @@
  * Sharable Link tooltip shows the view type (#2723)
  * New Cell buttons are faster (#2731)
 
+### Features
+ * Compatible with HTTP Basic Authentication for `github.com`. The option is `github.use.basic.auth`
+   and defaults `false` for compatibility with
+   [rcloud-gist-services](https://github.com/att/rcloud-gist-services)
+
 ### Bugfixes
  * `rcloud.download.file()` failed to insert newlines between
     character vector elements (#2721)